# batch/tassidicambio/target/generated-sources/openapi/src/main/java/com/example/restservice/api/ExchangeRatesFeesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.restservice.api;

import com.example.restservice.model.ExchangeRatesFeesRequest;
import com.example.restservice.model.ExchangeRatesFeesResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-05-09T14:26:53.298332800+02:00[Europe/Rome]")
@Validated
@Controller
@Tag(name = "Exchange Rates", description = "the Exchange Rates API")
public interface ExchangeRatesFeesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /exchangeRatesFees : Fornisce i costi relativi a commissioni di invio, tasso di cambio e commissioni di ricezione
     * 
     *
     * @param exchangeRatesFeesRequest  (optional)
     * @return successful operation (status code 200)
     *         or Default error sample response (status code 200)
     */
    @Operation(
        operationId = "getExchangeRatesFees",
        summary = "Fornisce i costi relativi a commissioni di invio, tasso di cambio e commissioni di ricezione",
        description = "",
        tags = { "Exchange Rates" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ExchangeRatesFeesResponse.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = ExchangeRatesFeesResponse.class))
            }),
            @ApiResponse(responseCode = "default", description = "Default error sample response")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/exchangeRatesFees",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }
    )
    
    default ResponseEntity<ExchangeRatesFeesResponse> getExchangeRatesFees(
        @Parameter(name = "ExchangeRatesFeesRequest", description = "") @Valid @RequestBody(required = false) ExchangeRatesFeesRequest exchangeRatesFeesRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"exchangeRatesFees\" : { \"sentFeesCurrencyTo\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"sentFeesCurrencyFrom\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"totalFeeCurrencyFrom\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"receivedFeesCurrencyTo\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"totalFeeCurrencyTo\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"feeExchangeRateCurrencyFrom\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"receivedFeesCurrencyFrom\" : { \"amount\" : 0.1, \"percent\" : 1.01 }, \"feeExchangeRateCurrencyTo\" : { \"amount\" : 0.1, \"percent\" : 1.01 } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/xml"))) {
                    String exampleString = "<exchangeRatesFeesResponse> </exchangeRatesFeesResponse>";
                    ApiUtil.setExampleResponse(request, "application/xml", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}